monde/pyproject.toml
[project]
name = "monde"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "glfw>=2.10.0",
    "numpy>=2.3.4",
    "pyopengl>=3.1.10",
    "pyopengl-accelerate>=3.1.10",
]
